(21) JDBC

+++ src/main/java/com/fawzy/springboot_fullstack/model/Customer.java
@@ -24,7 +24,7 @@ public class Customer {
             strategy = GenerationType.SEQUENCE,
             generator = "customer_id_seq"
     )
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private int age;
@@ -45,11 +45,11 @@ public class Customer {
 
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-//    public void setId(Integer id) {
+//    public void setId(Long id) {
 //        this.id = id;
 //    }
